Get rid of PNGExporter

+++ src/org/helioviewer/jhv/export/FFmpegExporter.java
@@ -9,38 +9,35 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.helioviewer.jhv.JHVGlobals;
-import org.helioviewer.jhv.Settings;
 import org.helioviewer.jhv.base.image.MappedImageFactory;
 
 class FFmpegExporter implements MovieExporter {
 
     private static final List<String> ffmpeg = List.of(new File(JHVGlobals.libCacheDir, "ffmpeg").getAbsolutePath());
 
-    private VideoFormat format = VideoFormat.H264;
-    private String path;
-    private int w;
-    private int h;
-    private int fps;
+    private final VideoFormat format;
+    private final String path;
+    private final int w;
+    private final int h;
+    private final int fps;
 
     private File tempFile;
 
-    @Override
-    public void open(String prefix, int _w, int _h, int _fps) throws Exception {
-        try {
-            format = VideoFormat.valueOf(Settings.getProperty("video.format"));
-        } catch (Exception ignore) {
-        }
-        path = prefix + format.extension;
-
+    FFmpegExporter(String prefix, VideoFormat _format, int _w, int _h, int _fps) {
+        path = prefix + _format.extension;
+        format = _format;
         w = _w;
         h = _h;
         fps = _fps;
-        tempFile = File.createTempFile("dump", null, JHVGlobals.exportCacheDir);
-        tempFile.deleteOnExit();
     }
 
     @Override
     public void encode(BufferedImage image) throws Exception {
+        if (tempFile == null) {
+            tempFile = File.createTempFile("dump", null, JHVGlobals.exportCacheDir);
+            tempFile.deleteOnExit();
+        }
+
         ByteBuffer data = MappedImageFactory.getByteBuffer(image).flip().limit(w * h * 3);
         try (FileChannel channel = new FileOutputStream(tempFile, true).getChannel()) {
             channel.write(data);
